Aabb support (#2)
* Submodule moment

* Create and update the bounding box every update

* Fix some errors

* Fix missing initializer valu

* Move bb info to data instead of instance and only calc on init

* Update libflipper submodule ref

* Fix vcpkg incompatibility for VS2022

* Revert "Update submodule"

This reverts commit 58c4e7a58fe2b8aeb177a6218dfd79bfcde6540f.

* Include <atomic> in a few places

Seems like a case of a platorm-dependent recursive include getting
what is needed from atomic on windows, but not on different implementations.

* Create and update the bounding box every update

* Fix some errors

* Fix missing initializer valu

* Move bb info to data instead of instance and only calc on init

* Fix vcpkg incompatibility for VS2022

* Include <atomic> in a few places

Seems like a case of a platorm-dependent recursive include getting
what is needed from atomic on windows, but not on different implementations.

* Type casting to fix tons of warnings

* Changes to make code more robust

* Create debug dir if not found

* Refactor and stabilize material code a bit. Still need to figure out UM  issue

* Small optimization

* Convert from broken png++ to stbi

* Small refactors

* Add padding field to properly align highlight member for ubo

---------

+++ src/J3D/Animation/J3DTexIndexAnimationInstance.cpp
@@ -72,5 +72,5 @@ void J3DAnimation::J3DTexIndexAnimationInstance::ApplyAnimation(std::shared_ptr<
 
     J3DAnimation::J3DTexIndexAnimationData data = *matFindResult;
     //material->AnimationTexIndices[i] = material->TevBlock->mTextureIndices[i];
-    material->AnimationTexIndices[data.TextureSlotIndex] = data.Track.GetValue(GetFrame());
+    material->AnimationTexIndices[data.TextureSlotIndex] = (uint16_t)data.Track.GetValue(GetFrame());
 }
